Add way to add global configuration to actions
Actions might require configuration options that are used across accounts,
like API keys and secrets. I added a GlobalConfigType method to the Action
interface. This method should return a struct instance of the type that will
hold the global configuration values.

I also added a GlobalConfig method to the ActionContext, that'll return an instance
of that struct, containing the current configuration values.